Codecleanup

+++ libraries/KustoLoco.Core/Evaluation/BuiltIns/Aggregates/MaxAggregate.cs
@@ -25,7 +25,7 @@ internal partial class MaxAggregate
     }
 
     internal static long? LongImplFinish(NumericAggregate context)
-        => context.Count == 0 ? null : (long)context.LongValue;
+        => context.Count == 0 ? null : context.LongValue;
 
     internal static double DoubleImpl(NumericAggregate context, double n)
     {
@@ -55,7 +55,7 @@ internal partial class MaxAggregate
     }
 
     internal static TimeSpan? TsImplFinish(NumericAggregate context)
-        => context.Count == 0 ? null : new TimeSpan((long)context.LongValue);
+        => context.Count == 0 ? null : new TimeSpan(context.LongValue);
 
     internal static DateTime DtImpl(NumericAggregate context, DateTime n)
     {
@@ -65,5 +65,5 @@ internal partial class MaxAggregate
     }
 
     internal static DateTime? DtImplFinish(NumericAggregate context)
-        => context.Count == 0 ? null : new DateTime((long)context.LongValue,DateTimeKind.Utc);
+        => context.Count == 0 ? null : new DateTime(context.LongValue,DateTimeKind.Utc);
 }
